commit quadratic equation addition changes

+++ src/Labs/lab5_quadraticEquation.java
@@ -28,8 +28,8 @@ public class lab5_quadraticEquation {
             x1 = (-b + Math.sqrt(D)) /(2 * a);
             double x2;
             x2 = (-b - Math.sqrt(D)) /(2 * a);
-            System.out.println("x1 = " + x1);
-            System.out.println("x2 = " + x2);
+            System.out.println("x1 = " + ((int) x1));
+            System.out.println("x2 = " + ((int)x2));
         }
     }
 }
